Add foreign key constraints

+++ dotnet/Sample.Database/dbo/Tables/State.sql
@@ -1,6 +1,6 @@
 ﻿CREATE TABLE [dbo].[State]
 (
-  [Id]          INT          NOT NULL IDENTITY,
+  [Id]          SMALLINT          NOT NULL IDENTITY,
   [Name]        VARCHAR(255),
   [StateCode]   CHAR(2),
   [DateCreated] DATETIME2(7) NOT NULL CONSTRAINT [DF_State_DateCreated] DEFAULT GETUTCDATE( )
